FilterByWidth analog to FilterHighway
