+++ app/process/bikelanes/bikelanes.lua
@@ -4,7 +4,7 @@ require("FilterTags")
 require("Metadata")
 require("HighwayClasses")
 require("RoadWidth")
-require("AddSkipInfoToHighways")
+require("FilterHighways")
 require("CheckDataWithinYears")
 require("StartsWith")
 require("categories")
