+++ app/process/shared/FilterByWidth.lua
@@ -1,0 +1,7 @@
+function FilterByWidth(tags, maxWidth)
+  local width = tonumber(tags.width)
+  if width and width < maxWidth then
+    return true, ";Skipped since `width<" .. maxWidth .. "` indicates a special interest path"
+  end
+  return false, ""
+end
