+++ app/process/shared/FilterHighways.lua
@@ -1,0 +1,41 @@
+function FilterHighways(tags)
+  -- Skip all non standard access values
+  local forbidden_accesses = Set({ "private", "no", "destination", "delivery", "permit" })
+  if tags.access and forbidden_accesses[tags.access] then
+    return true, "Skipped by `forbidden_accesses`"
+  end
+  if tags.operator == 'private' then
+    return true, "Skipped by `operator=private`"
+  end
+  if tags.foot == 'private' then
+    return true, "Skipped by `foot=private`"
+  end
+
+  if tags.indoor == 'yes' then
+    return true, "Skipped by `indoor=yes`"
+  end
+
+  if tags.informal == 'yes' then
+    return true, "Skipped by `informal=yes`"
+  end
+
+  if tags['mtb:scale'] then
+    return true, "Skipped since `mtb:scale` indicates a special interest path"
+  end
+
+  if tags.tracktype == "grade5" then
+    return true, "Skipped since `tracktype=grade5` indicates a special interest path"
+  end
+
+  -- Skip all unwanted `highway=service + service=<value>` values
+  -- The key can have random values, we mainly want to skip "driveway", "parking_aisle".
+  local forbidden_services = Set({ "alley", "drive-through", "emergency_access" })
+  if tags.service and not forbidden_services[tags.service] then
+    return true, "Skipped by `forbidden_services`"
+  end
+  if tags.man_made == 'pier' then
+    return true, "Skipped by `man_made=pier`"
+  end
+
+  return false, ""
+end
